Move from e2e/unit tests to tests
