+++ .travis.yml
@@ -14,11 +14,9 @@ install:
   - pip install -r requirements.txt
   - sudo ./setup.sh
 script:
-  - futurize --all-imports --stage1 --print-function --write --unicode-literals pwndbg
-  - git diff-index --quiet HEAD -- pwndbg
-  - isort --check-only --diff --recursive pwndbg e2e_tests unit_tests
-  - PWNDBG_E2E_TESTS_DISABLE_COLORS=yes nosetests -v -s ./e2e_tests/
-  - cd unit_tests/binaries && make && cd ../..
-  - gdb --silent --nx --nh ./unit_tests/binaries/reference-binary.out --command gdbinit.py --command unittests_launcher.py
+  - futurize --all-imports --stage1 --print-function --write --unicode-literals pwndbg tests
+  - git diff-index --quiet HEAD -- pwndbg tests
+  - isort --check-only --diff --recursive pwndbg tests
+  - PWNDBG_E2E_TESTS_DISABLE_COLORS=yes gdb --silent --nx --nh --command gdbinit.py --command unittests_launcher.py
   - python2.7 -m py_compile ida_script.py $(git ls-files 'pwndbg/*.py')
   - python3   -m py_compile               $(git ls-files 'pwndbg/*.py')
