--- e2e_tests/__init__.py
@@ -1,2 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
