+++ pwndbg/services/misc.py
@@ -1,1 +1,5 @@
+from __future__ import print_function
+from __future__ import division
+from __future__ import absolute_import
+from __future__ import unicode_literals
 import pwndbg.commands
