+++ pwndbg/services/misc.py.bak
@@ -1,0 +1,1 @@
+import pwndbg.commands
