+++ tests/binaries/reference-binary.c
