+++ tests/test_misc.py
